lab7

+++ app/Http/Controllers/CommentController.php
@@ -2,6 +2,9 @@
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
+
 class CommentController extends Controller
 {
     public function store(Request $request){
@@ -13,15 +16,17 @@ class CommentController extends Controller
         $comment->name = request('name');
         $comment->desc = request('desc'); 
         $comment->article_id = request('article_id');
-        $comment->user_id = 1;
+        $comment->user_id = Auth::id();
         if ($comment->save()) return redirect()->back()->with('status', 'Add new comment');
         else return redirect()->back()->with('status', 'Add failed');        
     }
     public function edit($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('update_comment', $comment);
         return view('comment.update', ['comment'=>$comment]);
     }
     public function update(Request $request, Comment $comment){
+        Gate::authorize('update_comment', $comment);
         $request->validate([
             'name'=>'required|min:4',
             'desc'=>'required|max:256'
@@ -32,6 +37,7 @@ class CommentController extends Controller
         else return redirect()->back()->with('status', 'Update failed'); 
     }
     public function destroy(Comment $comment){
+        Gate::authorize('update_comment', $comment);
         if ($comment->delete()) return redirect()->route('article.show', $comment->article_id)->with('status', 'Delete comment success');
         else return redirect()->route('article.show', $comment->article_id)->with('status', 'Delete comment failed');
     }
